[#132] Use query filtering for unapproved users, not result filtering
Unapproved users are not shown organisations and stakeholders. The DB
query returned these items, but these items were filtered in the handler
code. This made the pagination to behave weirdly, when the query
returned data that was being filtered out at the handler level.

This commit changes the DB query to explicitly request only for the
topic/resource types that are viewable by the user based on whether they
are an approved user or not.