[#110] Prevent unverified emails from being approved
We use Auth0 API and their machine-to-machine tokens to fetch the list
of users in the backend. Then we disable the "Approve" button for users
whose emails are not verified. We still allow disapproving of users
whose emails are not verified.

Auth0 provides us with 1000 tokens, for making API calls to fetch the
user list, for the tier we are in. For now, we just save this token in
an atom and re-use the token as long as it is valid. When the token
seems to be stale, we try to fetch a refreshed token from Auth0.

Closes #110