[#183] Add utility to email admins to notify new topics for approval
The utility function can be used to email all the (approved) admins,
when a new entity has been created that needs approval. Currently, the
utility isn't called anywhere, and needs to be hooked up.

The email sent is a simple plain-text email. It may be good enough,
since it is only being sent to the admins, but we could also make it a
nicer looking HTML email, if required.