[#190] Use country IDs instead of iso codes for showing country labels
In the browse end-point, the country field for all resources/topics
returns a country ID, except for the stakeholder, where we return an
iso_code. The iso_code doesn't seem to be used anywhere, other than
the detail page.

This commit moves from using the iso_code to get the country label to
using the ID. In a previous change, the iso_codes were made to be
non-unique because they aren't unique in the real world. AUS is used
for Australia and Ashmore & Cartier Is. (Aust.), for instance. This
commit fixes the bug of sometimes displaying the unintended country
label for a given country code.

The geo coverage values are still being returned as iso codes, and
there would surely be manifestations of this bug when showing the
country group information. But, that would require a more
wide-reaching fix which could come in a different commit/issue/PR.