[#190] Simplify code for toggling global geo coverage countries
Use an explicit flag to specify whether or not we want to display
list of all the countries when geo coverage type is global.

Also, due to a bug in the parser we parse the country values for
resources as geo coverage countries, even when the coverage type is
global. See Resource IDs 21 and 25 for example.

The UI was previously displaying these countries as geo coverage
countries, which we shouldn't. This commit changes the UI code to
ignore all the geo coverage countries when the type is global. The
full list of countries can be toggled using `showAllCountryList`.