[#110 #196] Remove :as json argument from auth0-util http client calls
The http library requires cheshire to do the json encoding and
decoding. The cheshire requirement is available in the dev env (as a
dependency of portal). But, is not available in the production env. We
could add cheshire as an explicit dependency, but it has some
requirements conflicts with jsonista. So, we go the route of using
jsonista to do the json encoding and decoding ourselves.