[#198] Add code to directly upload images to GCS
We directly upload images to the GCS bucket, instead of storing them
to the DB. The images are also directly served from the bucket, since
it is a public bucket.

"profile" and "event" type images still need to be served from the clj
server, until data from the DB is migrated to the Cloud Storage.

This commit doesn't change anything for "resource" type too. The
following commit will change that.