[#64] Reverse sort topics by ID in the browse view
Previously we sorted topics in the descending order of their creation
date, or their start date, but ascending order of IDs. This could make
some entities not appear where they are expected to be.

This commit changes the sort by ID to also be reverse, which is
consistent with the reverse sort by creation date.