[#225] Use backend type names in submission API data
We have a mess with the type names - stakeholder/profile/individual,
initiative/project, etc. Some of them are "front-end" names, and
others back-end names. Ideally, it would be nice to change the
back-end names to the front-end ones, but we should try to
consistently use them in the backend and front-end, and not mix the
two, until we make them uniform across the BE and FE. This commit
cleans up the submission API end-point to return the BE types
correctly, and simplifies the submission detail end-point.