[#224] Fix bug where country rows are updated twice
When updating the country ids, the same rows could potentially get
updated multiple times causing the IDs to multiple times, if the new
id that was already updated in the DB matches an old id that will is
replaced later in the sequence of replacements. This commit changes
the logic to keep track of already modified rows, and exclude them
when making future changes.