[#302] Don't send null geoCoverageValue when submitting event form
d37fd97826d796b676ba26217293ae554696211c set the geoCoverageValue to
`null` when the geoCoverageType changed, to clear any values selected,
when the type is changed to a different one. The API expects the
geoCoverageValue property to not be present in the request data if the
type is "global". But, this change caused a `null` value to be sent
for the geoCoverageValue, resulting in a 400 response from the server.

This commit deletes the geoCoverageValue property from the request
data if the geoCoverageType is "global".