[#132] Fix browse response for private topics for unapproved users
e62e610dceee07245bad4a7d3ff88a36d87caf46 changed filtering out private
topics for non-approved users to be filtering at the query level,
instead of querying the results. This was required to make the
pagination work correctly.

But, when a user queries for only a topic that they are not allowed to
query for, the response contains results for all the allowed topics
because the SQL query ignores filtering by topic.

This commit fixes that by setting the list of allowed topics to an
empty array, and the returned result contains 0 entries.