[#184] Use counts returned by browse query instead of landing counts
The previous counts were a hack, where we tried to use the counts in
the landing page for the counts in the browse view. This commit uses
the count values that are now returned by the /browse API call. This
also fixes the pagination to be correct for different searches.