[#246] Correctly tags and urls when editing a resource
The form doesn't send the IDs of the *_tag row, or *_language_url
row. We only have the list of tags or URLs to use for the resource.

The way we handle the edit right now is to delete all the related
data, and recreate it. Potentially, this can be improved to delete
only what has been removed, and add what has been created.