[#326] Make /landing & /browse count match
We have decided as per issue #72 to remove the 'global' resources when
filtering country specific data.  One of the user facing problems is
that some countries didn't have any "Initiative" happening there and
still the counts where showing something different than zero.

When using the /browse to search, by default all the 'global'
resources where showing up, instead of the country specific ones.
IMHO there needs to be a decision per resource type to include those
global resources, but sort them at the end of the list. Something
like: Country > Region > Global - (Note: Something to be discussed and
agreed first)

This changes only make the counts consistent between `/landing` and
the results on `/browse`. Doing the filter on 'regional' resources
implies that we need to change all the queries per resource type, and
come up with a notion to define a value for 'regional'. Right now, `0`
means 'global'

This changes are not the final solution, but a more consistent one.