[#283] Show Other option when typing names not in search list
The sign up form shows the Other option when the name searched for is
not present in the list of options. This commit makes the other forms
behave similarly, when the Select list has the option added to the
list of options with a value -1. The organisation lists in the forms
already have this option added.