[#363] Add topics to the DB
Ideally, the tags would've been added to the tags.json file so that the seeding
process wouldn't overwrite these tags, but tags don't seem to have IDs as of
now, and the seeder doesn't update the tags when it is re-run. This could be
changed in a future commit, later.