[#369] Add end-point to update the role of a user
This commit adds a PATCH end-point that currently only allows updating
the role of a user. This method can be extended to update other
attributes of the user. The API params can be split into multiple OR'd
maps allowing groups of parameters to ensure a sane subset of
parameters, each time a PATCH request is made.