[#685] Refactor topics query to use original query insteal of views
The query can either count the topics or query their data depending
on given parameters. This way, if we ever need to change the topics
query we don't need to consider all the existing views or look back in
migration files. We have it in a single place.

Next steps will be to abstract the repetitive parts of the query to
reduce it's visual size.