[#589] Get `geo_coverage_values` back
This was commented earlier to allow bypassing entering these values
to the db as there was some issue with them, but this instead resulted
in the pipelines failing because of them being linked in several
places.
However, in the end, it was concluded that the issue regarding this
whole topic could be resolved from the FE instead, so this commit
uncomments everything and brings them back the way they were.