[#651] Add startDate and endDate parameters API spec
* I also added ^:const meta to previous exisiting constants. Constants
should always have this meta so at the runtime unecessary steps such
as Var.getRawRoot can be skiped which otherwise would add a bit more
of latency. At any rate, performance is not an issue here of course,
this is just to settle a convention to adopt the usage of ^:const when
necessary.