Revert "[#60] Added rolling strategy"
This reverts commit 01ace6e96cc808249190f249eb7667638aad7cd6.

I'm not yet quite sure why, but deploying to production has recently been failing.
The solution has been to delete the new pod, which somehow leads K8 to believe
 that the rolling update is done.
It still starts a new pod because the new revision demands it and the old one
 is removed.
When the new pod is the only one starting, then the update goes over smoothly.

At the moment, I cannot explain why this is happening only in production and not on test.
But since it looks like running a singular pod is the solution, hopefully,
 this'll resolve our issue.

Having an old version writing to the DB during or after a migration,
 probably doesn't leave the application in a good state either,
 so this might be cleanest way to start a new version.