[#1145] Improve prod Dockerfile JVM startup configuration
* This will print out the JVM configuration on start up.
* It adjusts the JVM memory heap allocation to respect the imposed
memory limit by the app's container.
* MaxRAMPercentage set the percentage total amount of memory available
for the VM.
* It splits the initialization of the migrations and the main
application's process. We want to execute the migrations first and if
it fails to execute them we want the application to crash. Failing
silently could be a problem later on. So it first executes the
migrations in a JVM process and when it finishes another process
starts to start the application's main process.