Remove storage of local deps
The backend starts up really slowly on Macs (20-30 minutes).
Part of the reason are the volumes which are known to have terrible performance.

In order to make it liveable for developers on macs,
 the deps are moved into the image.
The virtual machine that runs the container will thus only have to share the
 backend code on the host.
The deps will be available in the virtual machine with less overhead.