Add image and resource type in Get Detail API
[Closes #1184]

Included resource type info for the related content data loaded for a resource
retrieved using GET Detail API endpoint. In this case, we hardcode the
resource type for all the resource involved that don't have its own "type".
However, for the "Resource" table we use the "type" property, since this is a
dynamic info. In any case, we provide the "type" value in lowercase and
replacing spaces by underscores, so FE can use this property for navigation
when visiting a related content from a given resource.

Besides, we also load requested image data for each related content. In the
case of "project" (initiative) resource type, we keep using "image" property
to return a consistent response (as in this case the property is
called "qimage").