Stop generating default stakeholder avatars in BE
[Re #1256]

Now we always try to use the provided user image as avatar, setting it to null
if there is not any, as before we were generating a default one, which is not
desired anymore.

Updated tests that were expecting the default avatar image to be created
(also the deleted tests were duplicated).