Remove default stakeholder avatars in DB
[Closes #1256]

Deleted (set to null) stakelholder avatar pictures that were detected as
defaults. For this operation we consider two possible picture URL patterns as
one of the patterns is not used anymore, but we can still have stakeholders
using it.

As a side note, this is the first only-up migration we write, since we were
not sure until now if we could just skip the down (rollback) migration, since
we are not using them because of the raise-error strategy, so it does not
make sense to keep writing them, now that we know we can do it in this way.
Besides, in this case there is no going back for this operation.