Add image to organization's related content info
[Closes #1280]

Added image property for the query to retrieve each related resource belonging
to an organization. The relationship can be set directly to the organization
or an stakeholder belonging to a given organization, but we cover both cases.
This allows FE to display the image when viewing the related resources of an
organization.

We keep using same "image" property name returned by the query in order to
keep consistence and be able to use this as stable/expected property name
regardless of the resource type.