1145 improve application logs and configuration (#1225)
* [#1145] Improve prod Dockerfile JVM startup configuration

* This will print out the JVM configuration on start up.
* It adjusts the JVM memory heap allocation to respect the imposed
memory limit by the app's container.
* MaxRAMPercentage set the percentage total amount of memory available
for the VM.
* It splits the initialization of the migrations and the main
application's process. We want to execute the migrations first and if
it fails to execute them we want the application to crash. Failing
silently could be a problem later on. So it first executes the
migrations in a JVM process and when it finishes another process
starts to start the application's main process.

* [#1145] Set explicit configuration for ragtime

* These option are the default but it is better to have them explicit
in the configuration rather than going to the docs/source-code to know it.

* [#1145] Set prod log level to :warn

* I need to investigate but the :min-level is ignored for some
reason. After talking internally with Iñaki the :set-root-config?
might be the culprit.
* Setting the :level explicitly solves the problem and we won't have
:info level log entries in production.
* Next steps will be to introduce a middleware for timbre to
dynamically change the log configuration while the application is
running. It will surely be useful to debug things in
production. However, the application barely has logs, so one
improvement would be introducing logs where it's useful.

* [#1145] Increase the timemout for wait4ports

* DEBUGGING: With the new strategy to run migrations and main application process
separatly, the backend server can take a bit more of time.

* [#1145] Revert log level to :info for production config

* [#1145] Re-add :set-root-config? in timbre config

* I accidentaly removed this configuration.

* [#1145] Disable logger for hikaricp to not log SQL entries

* nil doesn't work because Duct will prefer a non nil value over nil
when merging configurations.