Remove legacy policy views and tables
[Re #1273]

Removed legacy policy views that were migrated to actual DB queries, as they
are no longer used and we need to get rid of them in order to make simpler
the process of updating some columns of `policy` table or being able to
delete the now legacy `policy_language_url` table (something we need for
performing the LEAP API integration).

As we are deleting `policy_language_url` table (since this is no longer used),
we have also updated BE code that were sending or expecting this column.

For now we have only removed code related to "policy" entity, but we should
make a bigger refactor later to get rid of this code in other resources that
no longer use that multilingual url feature.