Erase policies and adapt DB model for LEAP API
[Re #1273]

Deleted all the policies and related data from the current DB via migration,
since we are going to export all of them from LEAP API and current existing
policies in the app are got from there.

Modified structure of "policy" table to use date format
for "first_publication_date" and "latest_amendment_date", since those should
be the format set from the beginning for these columns and it is what it
matches the LEAP API respective fields. In this way, we avoid lot of
complexity handling time zones, which is not part at all of these fields.
Also we are providing this format already for FE, so this change is aligned
with that fact too.

Adapted code to handle new date format for mentioned fields. In this case, we
have needed to modify a generic resource update function to take into account
this special case for "policy" entity type, since we were assuming we could
convert the columns into timestamps for all the resource types, which is no
longer the case. In the future we should probably refactor the code to keep
same column types along all the resource types, since they are probably meant
to use dates here.