[#1317] Update stakeholder schema for experts
* Add new expertise tags and categories.
* Add new invitation table (refactored the old one as it was legacy
and tied to organisations).
* Remove representation column from stakeholders as it's not used
anymore. As a side-effect I had to remove the views v_stakeholder,
v_stakeholder_data, v_topic and v_topic_all because they depended on
that column and because we are going to remove views anyway. We should
avoid views until the model is consolidated. Other views will be
removed in further iterations.