Refactor create tags db fn to accept any column
[Re #1273]

As we needed to re-use `new-tags` db fn for tags creation allowing
`review_status` column to be set for LEAP API import operation, we have
refactored the mentioned function to receive columns dynamically as
parameter, so we can use it in a more flexible way and also cover our
specific needs.

Refactored `new-tags` db function to use the verbose backend convention to
describe the command and output of the query.

As such, we have refactored the code that was using the function above,
providing explictly the columns as parameter.

In order to ensure the order of insert keys and values is respected, we have
extended the sql utility function that generate the insert columns based on
the entity cols to also allow insert keys to be passed directly, so we can
ensure the order in that way, since the values are generated separately and
it could be dangerous otherwise.