Add support for tags creation in Leap API import
[Re #1273]

Added support to process `keywords` property of the policies to be imported as
tags, where we create them if they don't exist already.

We have refactored the code to avoid additional queries to get the updated
list of existing tags each time we process a batch, by sharing the new tags
as part of the import result, merging the new tags with the ones provided at
the beginning.

Modified migration related to LEAP API integration's needed changes to include
a new `tag_category` ("leap api"), specific for being used to relate the new
tags created from LEAP API. In this way, we can clearly distinguish them for
different purposes and we don't rely on dynamic data that could be different
in each environment.

Refactored `create-policies` db function to return `leap_api_id` for each row
inserted, so we can use this info later to relate the generated entity ids
with their `leap_api_id` identifiers from LEAP API.

Added missing docstrings to some pending sql utility functions, now that we
have extended one of them, as it is important that the behaviour is clear.