Support update of tags from Leap API
[Re #1273]

Added update support for incoming tags from LEAP API, that have the shape of
merging different properties. For now we only use `keywords` property but we
will extend the support to adding other tags from other properties.

We only add new tags but we do not remove any existing one, neither touch the
existing policy-tag relationships, since they could be edited from the
platform and we want to avoid conflicts. In that way, any tag coming from
Leap API will be able to be overriden by Unep Platform admins, if needed.

We also handle duplicate of tags, so we skip introducing them in different
stages of the process. This is important, since we don't have right now any
unique constraint to ensure that each row of policy-tag table has a unique
combination of policy and tag ids. This is on process to be changed, but for
the time being we need to handle this situation.

Tags duplication is also avoided for the "tag" entity since we do not have
unique constraint for their name. In our case we will pick the first
available tag under a given name index.

For the check of existing policy-tags we have created a new DB query as the
current generic resource tag select query was not suitable for the needs and
it was limited to get approved tags only. Also the information needed is
quite less for this use-case and the format/naming we give to each property
can be controlled better in this query, keeping the solution simpler by using
a separated query for this purpose.

Removed outdated comments as well.