Support "plasticToolkitTags" field from Leap API
[Re #1273]

Now we have introduced support for parsing the mentioned field from Leap API
incoming policies, so it can be persisted when creating or updating those
policies.

At the same time, the parsing function to generate tags now work combining any
additional tag-containing properties, so we could easily support additional
tag-related properties in the future. In this case, we make sure we only
introduce distinct tags as well.

Fixed a docstring format as well.