Handle geo-cov. data from created Leap API pol.
[Re #1273]

Now we use the country information to also populate "policy_geo_coverage"
related data when importing policies from Leap API for creation process.

We still need to support the Update process.

Refactored existing DB query to be flexible when accepting insert columns.
Also the name and expected return should be plural as we are using it for
multiple policies from Leap API usage and that can be used for single usage
as well. The calling functions have been adapted to this change as well,
hardcoding the columns as it was the easiest way to do it.