Handle geo-cov. data from updated Leap API pol.
[Re #1273]

Added support for Update operations, using same strategy as for creation. In
this case, we delete previous records of policy-geo-coverage rows before
re-creating all of them with updated info. We do it in that way since it is
the simplest way, as Leap API will be the only source of truth for this
field.

We have added a new DB query for the deletion process, using policy ids as
targeting condition, since that is what we have from the Leap API and should
be good enough.