Basic support for implementingMeas from Leap API
[Re #1273]

Added support for this property by storing the first incoming value from the
list, when available.

In next iterations we will refactor this to store multiple values, as the
model will need also to be changed.

Added a TODO to improve how we look into the country group options.

Removed an outdated comment.