Schedule Leap API initial import
[Re #1273]

Added a default env var value for scheduler to work without the need to add
those env vars into GCloud, as a temporary solution until we can create those
env vars.

The idea is to set the initial scheduling when the deployment to test env is
going to happen and then change it again for Prod env, once we know it is
working fine in test env. After that, we can schedule it each 2 weeks, as
expected.

The timezone default settings has been also set for convinience but that can
be later changed anyway.

Scheduler Thread Count setting is meant to provide a maximum of threads
available for the scheduler, that can be changed in any moment. The current
value is something we have used in other projects.