Be bugs (#1384)
* Fix get-stakeholder-associated-topics to return topic/resourcce type

* We weren't returning the resource/topic type. Only for a resource
that has the `type` property in its table by chance.
* I also took the opportunity to refactor a bit the namespace:
integrant keys to be explicit, and API opts for parsing to use a conditional update.

* Fix redundant fn wrapper in invitation API

* Fix invitations also to return details about the stakeholder

* Return first_name, last_name, and tags unwrapped.
* I also had to tweak the `get-reviewers` handler because the
`get-stakeholders` query also changed. It assumed review_status
`APPROVED` always, but it should be more generic and allow passing the
`review_status` as a filter.