Fix popular topcis counting of Browse API GET endpoint
[Closes #1404]

The counting logic was modified for ignoring duplicate tags, but it has not
working well, since all the resources were counted for each popular topic
(tag). Now we have fixed that part of the DB query to do the right counting,
ignoring duplicate tags until the root problem is fixed.