Add `featured` prop to Resources
[Re #1406]

Added new property in the DB model as specified in the issue, using a default
`false` value for it.

Just because we have added it, we get it "for free" also in Browse API Get
endpoints and the ones related to Resource Detail view, which will be useful
for next steps (specifically for Browse API where we want to use it).