Add resources' `featured` in GET Submissions API
[Re #1406]

Added the property in the DB query to fetch the submissions as part of the
`data` property from the response. In order to do that, we have added the
property to be got from the resources but we have also needed to include it
as NULL, casted to BOOLEAN, for the rest of the submissions data entities, in
order to be able to do the UNION operation over all the entities. The casting
was needed because Postgres seems to treat NULL as text, for some reason,
which makes the UNION to break, as we need same columns with same types to
make it work.

We are leaving the new property as NULL for all the submissions, but that does
not harm when we are not dealing with Resources.