Allow `featured` prop in PUT Submission API
[Re #1406]

Allowed this endpoint to pass an optional `featured` param to indicate that a
resource needs to be updated in the submission to consider it as featured
resource.

Beware that this new param should not be passed for other entities that are
not resources, since the operation would fail, as this flag is only available
for resources.