Allow filter by `featured` in Browse GET API
[Closes #1406]

Now we are able to filter results and count of resources by their `featured`
property, by allowing an extra optional param in the API with same name.

Left a TODO for extending the tests, since we need this ASAP. We can take care
of that when we add sorting capabilities in further iterations.