Fix duplicated tags
[Re #1416]

Fixed duplicated tags by DB migration, where we keep the relationships between
tags and related resources, before getting rid of both duplicate
relationships and tags.

For that we use a couple of views in order to re-use duplicated tags results
needed for the migration process, to know which tags relationships need to be
changed and then delete duplicated relationships and tags. After the process
is done we get rid of the views as well.

Once duplicated tags and relationships are gone, we ensure it cannot happen
again, by adding some constraints to the respective tables.

All the additional details have been documented in the migration file.