Ensure tags are stored as user types them
[Closes #1416]

We have needed to change the way we store and check for existing tags in order
to let be stored as they come, without lowercasing them. Related to this
change, we expect FE to send the existing tags as BE has stored them. There
is a problem here, since FE had some hardcoded names that do not match with
what BE has stored (as previously we were lowercasing all tags before storing
them). But FE will solve this by using what BE has stored and sent to FE.
Thus we make sure the tags appear in a single way all the places are
displayed and the handle of tags is kept simpler.

On the other hand, as tags need to be unique, we have needed to handle any
exception related to this, since it could happen that due to concurrent
requests some tags are attempted to be re-created. In order to achieve this,
we have needed to provide logger to several API handlers related to
resource/stakeholder/org creation/edition, so we can log this kind of error,
while error handling can be done in different ways at different steps.

When creating stakeholders now we return a bad request error with the details
if we detect it happens because of duplicated tags attempts. In the future
this error handling should be extended but it was too much work to do it
now.