Fix policy languages and unify for translations
[Re #1422]

As we currently do not support multi-lingual feature for imported policies
from LEAP API, we have fixed the set language for existing ones to use
English as the original language, since we can only ensure the content to be
available in that language (ignoring some bug from LEAP API where we get the
content marked in english but in wrong language).

As such, we also hardcode the language of incoming imported policies from LEAP
API to use the english as language, since we are getting and expecting all
the content in english (we still store original title but we do not have yet
enough info neither resources to store translations for imported policies).

Besides, we make sure now that we require language (as the "iso_code") in
order to create new policies or update them. At this point, we have get rid
of the way we were updating language from policies, as something aside, since
now we require this property to be present all the time and it can be
directly set without any special parsing.

Changed the query to add policy language info when Getting a policy detail, in
order to use the update reference. However, we will probably not need this
extra step if we just want the language info to use it for the updates.

As a final note, we are creating english language entity if it's not present
as part of the migration to fix and unify policy language data, since we need
to enforce a default language to existing items and this language is not
present when executing tests on local DB for tests. As such, we have removed
this language from the imported data set for the Seeder tool, since we are
already creating it in the DB migration.

Updated some tests related to this change, but Seeder-used data is too
outdated to start changing that now. We should review that part soon.