Support language property in remaining resources
[Re #1422]

We have added `language` property (DB model) to the remaining resource types,
in the same format as we have done for the Policy type. In this case, the
operation was a bit different, since it is a new property, so we have added
the same english default value for existing resources in the DB, but dropping
that default after updating existing rows, making mandatory the new property,
as with Policies.

Updated handler and DB queries related code to take this new property into
consideration. As such, now we get this new property when requesting a
resource and we require it when they are created. However, we do not require
the property when updating the resource, since we use a generic endpoint
which includes organizations and stakeholders, that do not include this
property. Also, it can happen we just do not want to update this field but
others.

We have needed to also modify the Seeder and testing tools to include this new
property without the need to touch all resource data files by hand, as that
would be too much work for nothing. Thus, tests are also updated in this
regard.

As a side note, we have not touched SQL queries that are no longer used, to
optimize efforts.