Add resource translation tables (DB model)
[Re #1422]

Added translation tables for all the resource types, as it has been
specified.

With this DB model design, we can easily do upserts for updating or creating
translations for a given resource, field and language. In order to manage the
translations, we just need to make sure the `translatable_field` values are
defined in the Domain layer, so only expected/allowed values are used
(BE needs to control that and make sure lowercased values are always used).
We could also use some enum for the `translatable_field` property value, but
that would make harder the maintenance, since enum values cannot be just
deleted and in this way we can update an work with translations in a more
agile way.

Notice that we also have constraints to update or delete on cascade the
translations if any referenced language or resource has been modified or
deleted. This reduces manual work and improves integrity.