Implement Res. Translations PUT endpoint
[Re #1422]

Added new endpoint for creating and editing any kind of resource translations,
as specified.

The limitation here is that all the translation need to belong to the same
resource, but we can create or edit them using the same endpoint.

For managing the permissions we have extracted the code from
the "handler.detail" related namespace, but fixing the bug it was there,
since we were checking wrongly the permissions added to the target topic
(we were not doing the comparison properly, as the result was in a sequence).
However, we have put the modified code in a shared space, without touching
the original code for managing Resource Detail related endpoints permissions,
not to do all at the same time. We will refactor the code to use the new
fixed shared version soon.

Added some tests for the new endpoint, but leaving some cases, as we are not
considering fully the params spec part here.

As a side note, in order to verify the fields that can be used for
translations, we have needed to receive the same `topic-type` param also from
the request body, but still makes sense to receive it in the URL as well, to
keep consistency.