Implement Res. Translations GET endpoint
[Re #1422]

Added new endpoint for getting either available resource translation languages
or directly get available translations, as specified.

The limitation here is that all the translation need to belong to the same
resource, but this all can be extended in the future, among other filtering.

For managing the permissions we are using similar approach as for the Get
Detail API endpoint, using shared resource permissions management namespace.

The grouping we do is for help FE to manage the translations in a more
easier/natural way.