Add Res. Translation Delete endpoint
[Re #1422]

Added pending endpoint to delete translations for a given resource, using
different filters available.

We should add automatic tests when possible, for the next iterations.

Besides, we should improve documentation in Swagger, regarding the options for
filtering which resource translations can be deleted.

Not closing the issue yet, in case we need to do further changes.